levelup_habits/habits/serializers.py
from rest_framework import serializers
from .models import Habit, Completion

class HabitSerializer(serializers.ModelSerializer):
    class Meta:
        model = Habit
        fields = '__all__'
        read_only_fields = ['user', 'created_at']
        
class CompletionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Completion
        fields = '__all__'